refactor(api): extract article analysis profiles
